queryRootResolverをmainから分離
